fix: queueing a job with a duplicate id triggers a json error
If you try to queue a job with an id that's a duplicate of a job
id already queued (say if your job ids are deterministic, and some
condition resulted in one being queued twice), the duplicate won't
get requeued, which is reasonable. This means that the `fetch_one`
query at the end of add_job() finds nothing, resulting eventually in
responding to the caller with a json decoding error.

The fix is to fetch_optional and respond with the option, which ripples
a signature change through several functions.

This fix is a BREAKING change in a very important API, so you might
not want it. The alternative would be to add an equivalent alongside
it that has the optional behavior.